pushing several scenes for Lui to add sounds too
going back in scenes does not work in this version

+++ src/app.js
@@ -1,3 +1,4 @@
+var INITIALIZED = false;
 
 var GameScene = cc.Layer.extend({
 	ctor:function() {
@@ -11,9 +12,16 @@ var GameScene = cc.Layer.extend({
 var Runner = cc.Scene.extend({
 	onEnter:function () {
 		this._super();
-		localStorage.clear();
-		var layer = new GameScene();
-		this.addChild(layer);
+		if(INITIALIZED == false) {
+			
+			localStorage.clear();
+
+			INITIALIZED = true;
+			
+			var layer = new GameScene();
+			this.addChild(layer);
+		}
 	}
 });
+
